DDF-UI-248 use other polyfill and cleanup
