+++ ui/packages/catalog-ui-search/package.json
@@ -65,6 +65,7 @@
     "density-clustering": "1.3.0",
     "dropzone": "4.3.0",
     "eonasdan-bootstrap-datetimepicker": "Eonasdan/bootstrap-datetimepicker#4.17.47",
+    "eventsource": "^1.0.7",
     "font-awesome": "4.7",
     "geo-convex-hull": "1.2.0",
     "golden-layout": "1.5.8",
